email/python/requirements.txt
# Google API dependencies
google-auth>=2.0.0
google-auth-oauthlib>=0.8.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0

# Microsoft Authentication Library
msal

# PDF processing
pikepdf>=6.0.0

# HTTP requests (if needed for any additional features)
requests>=2.25.0

# PDF & HTML parsing
PyMuPDF
pytesseract
pillow
beautifulsoup4
lxml
html5lib